Moved method get_user_residential_address into main::RetrieveUserResidentialAddressMixin class
